% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{addGroupName}
\alias{addGroupName}
\title{addGroupName}
\usage{
addGroupName(libs, samples)
}
\description{
Displays something like "samplename | groupname".
}
\details{
Takes a vector of sample names name and constructs a vector of
strings made of the sample name, a pipe separator and the name
of its group.
}
