% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{addNumberOfSamples}
\alias{addNumberOfSamples}
\title{Add number to group names}
\usage{
addNumberOfSamples(libs, groups)
}
\description{
Displays something like "groupname | n".
}
\details{
Takes a vector of group names name and constructs a vector of
strings made of the group name, a pipe separator and the number
of samples in the group.
}
