% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{contentgenesPCA_1}
\alias{contentgenesPCA_1}
\title{contentgenesPCA}
\usage{
contentgenesPCA_1(subgenes)
}
\arguments{
\item{df}{An expression table (see above)}
}
\description{
Vizection's Principal Component Analysis
}
\details{
First, the expression table, assumed to contain counts,
is normalised to tags per million (TPM) and transposed.
Then, the first line, containing the combined expression
of all the CTSS that were not contained in an annotated
feature (such as a gene), is removed.  Then the PCA is
run with the ade4 package's function dudi.pca().
}
\examples{
vizection:::vizectionExampleGenes() \%>\% contentgenesPCA

}
