% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalClustering.R
\name{contentheatmapGenes}
\alias{contentheatmapGenes}
\title{Plots Vizection's heatmap of correlations.}
\usage{
contentheatmapGenes(cormat, dendr, sublibs)
}
\arguments{
\item{cormat}{A correlation matrix like the output of corMat().}

\item{dendr}{A dendrogram object like the ouptut of gendsDend2().}

\item{sublibs}{A "libs" table like the output of vizectionExampleLibs().}
}
\description{
Plots Vizection's heatmap of correlations.
}
\examples{
cormat <- vizectionExampleGenes() \%>\% corMat
dendr <- cormat \%>\% distCorMat \%>\% genesDend \%>\% genesDend2(x = vizectionExampleEnv())
contentheatmapGenes(cormat, dendr, vizectionExampleLibs())
}
