% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{filterSelectionBool}
\alias{filterSelectionBool}
\title{Apply filters to samples}
\usage{
filterSelectionBool(libs, input)
}
\description{
Apply filters to samples
}
\examples{
filterSelectionBool( libs = vizectionExampleLibs()
                   , input = vizectionExampleEnv())
}
