% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalClustering.R
\name{genesDend}
\alias{genesDend}
\title{genesDend}
\usage{
genesDend(d)
}
\arguments{
\item{d}{a distance matrix}
}
\description{
Cluster a distance matrix with the complete method.
}
\examples{
data.frame(1:3, 2:4, 6:4) \%>\% corMat \%>\% distCorMat \%>\% genesDend
}
