% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalClustering.R
\name{genesDend2_1}
\alias{genesDend2_1}
\title{genesDend2}
\usage{
genesDend2_1(x)
}
\arguments{
\item{x}{A list where the element "nbClusters" is the number of
clusters to compute and "nbGroups" is the number of
labeled groups.}

\item{d}{a distance matrix}
}
\description{
Cluster a distance matrix with the complete method.
}
