% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pcaCompGenesList}
\alias{pcaCompGenesList}
\title{pcaCompGenesList}
\usage{
pcaCompGenesList(components, n)
}
\arguments{
\item{components}{Principal components of a PCA object of class
"dudi", extracted by subsetting it for its "co" member.}

\item{n}{Rank number of the principal component to select}
}
\value{
The "components" data frame, sorted on the "n"th component.
}
\description{
It's complicated...
}
