% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotEigenValues}
\alias{plotEigenValues}
\title{plotEigenValues}
\usage{
plotEigenValues(orderedCompPca)
}
\arguments{
\item{orderedCompPca}{a The PCA object produced by contentgenesPCA.}
}
\description{
Plot the Eigen values of the "dudi" PCA object produced by contentgenesPCA.
}
\examples{
vizection:::vizectionExampleGenes() \%>\%
  contentgenesPCA \%>\%
  plotEigenValues
  
}
