% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotHTB}
\alias{plotHTB}
\title{plotHTB}
\usage{
plotHTB(orderedCompPca, comp, nbDispGenes = 25)
}
\arguments{
\item{orderedCompPca}{A data frame where columns are principal component,
the rows are genes or similar features, and the component to
plot is already sorted.}

\item{comp}{The number of the principal component in that matrix.}

\item{nbDispGenes}{How many bars for the head and the tail (each).}
}
\description{
plot head and tail as barplot.
}
