% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{UNaddNumberOfSamplesOrGroup}
\alias{UNaddNumberOfSamplesOrGroup}
\title{Clear sample or group names.}
\usage{
UNaddNumberOfSamplesOrGroup(names)
}
\arguments{
\item{names}{Group or sample names to which other information
have been added by the functions addGroupName or
addNumberOfSamples.}
}
\description{
Same as above except that it does not keep name attributes.
}
\examples{
c("toto | 5", "H12 | toto") \%>\% UNaddNumberOfSamplesOrGroup
}
