% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{vizectionExampleGenes}
\alias{vizectionExampleGenes}
\title{vizection example \dQuote{genes} table}
\usage{
vizectionExampleGenes()
}
\description{
Toy example for the \dQuote{genes} table to be used in vizection.
}
\details{
This is for debugging purposes; the example does not reflect well
the kind of real data that is expected.
}
\examples{
summary(vizectionExampleGenes())
head(vizectionExampleGenes())
}
