% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{vizectionExampleLibs}
\alias{vizectionExampleLibs}
\title{vizectionExampleLibs}
\usage{
vizectionExampleLibs()
}
\description{
Toy example for the "libs" table to be used in vizection.
}
\details{
This is for debugging purposes; the example does not reflect well
the kind of real data that is expected.
}
