% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizectionSummarizedExperiment_Class.R
\docType{class}
\name{VizectionSummarizedExperiment}
\alias{VizectionSummarizedExperiment}
\title{VizectionSummarizedExperiment, S4 class that extends \code{"SummarizedExperiment"} class.}
\description{
Extension is to add the results of the analyses runned in Vizection as well
as recording session version (new feature to come).
}
\section{Slots}{

\describe{
\item{\code{dendrogram}}{The dendrogram computed in vizection.}

\item{\code{pca}}{A principal components analysis runned with \code{ade4} package.}

\item{\code{coa}}{A correspondance analysis runned with \code{ade4} package.}

\item{\code{vizectionParams}}{List of parameters used as metadata for running Vizection sessions.}
}}

