% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionValidate.R
\name{vizectionValidate}
\alias{vizectionValidate}
\title{Validate objects passed to Vizection}
\usage{
vizectionValidate(genes, libs)
}
\arguments{
\item{genes}{A \sQuote{genes} table.}

\item{libs}{A \sQuote{libs} table.}
}
\description{
Makes an error with objects that do not fit the requirements for being
used in Vizection
}
\examples{
# Make arbitrary input data using the "iris table".
ilibs <- iris
ilibs$group <- ilibs$Species
ilibs$samplename <- paste("sample", 1:150)
rownames(ilibs) <- ilibs$samplename
ilibs$Run <- "Run1"
ilibs$counts <- 10

igenes <- t(iris[,1:4]) \%>\% data.frame
colnames(igenes) <- rownames(ilibs)

# Check that the format is valid (silently returns TRUE).
vizectionValidate(genes=igenes, libs=ilibs)
}
