% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalClustering.R
\name{genesDend2_1}
\alias{genesDend2_1}
\title{Cluster a distance matrix with the complete method.}
\usage{
genesDend2_1(x)
}
\arguments{
\item{x}{A list where the element "nbClusters" is the number of
clusters to compute and "nbGroups" is the number of
labeled groups.}

\item{d}{a distance matrix}
}
\description{
Cluster a distance matrix with the complete method.
}
\examples{
dendr <- data.frame(1:3, 2:4, 6:4) \%>\% corMat \%>\% distCorMat \%>\% genesDend()
genesDend2(dendr, x = vizectionExampleEnv())
genesDend2(dendr, x = vizectionExampleEnv() \%>\% inset("showGroupsColor", FALSE))
  
}
