% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{vizection}
\alias{vizection}
\title{Vizection}
\usage{
vizection(genes = genes, libs = libs, local = F)
}
\arguments{
\item{local}{T if you want host to be "0.0.0.0"}
}
\description{
This function launches the Vizection shiny application.
}
\details{
By default, Vizection will look for two data frames called \dQuote{genes}
and \dQuote{libs}.  Alternatively, these data frames can be passed as
arguments to the \code{vizection()} function.  Their names will be used to
update the global options (\code{vizection.genes} and \code{vizection.libs})
that Vizection uses to find the objects in the global environment.
}
\examples{
vizection(genes, libs, F)
}
