% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{vizection}
\alias{vizection}
\title{Vizection}
\usage{
vizection(genes = genes, libs = libs, local = F)
}
\arguments{
\item{local}{T if you want host to be "0.0.0.0"}
}
\description{
This function launches the Vizection shiny application.
}
\examples{
vizection(genes, libs, F)
}

