% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{vizection}
\alias{vizection}
\title{Launch the Vizection shiny application.}
\usage{
vizection(se, genes, libs, local = F)
}
\arguments{
\item{local}{T if you want host to be "0.0.0.0"}
}
\description{
By default, Vizection will look for two data frames called \dQuote{genes}
and \dQuote{libs}.  Alternatively, these data frames can be passed as
arguments to the \code{vizection()} function.  Their names will be used to
update the global options (\code{vizection.genes} and \code{vizection.libs})
that Vizection uses to find the objects in the global environment.
}
\examples{
\dontrun{
library("airway")
data("airway")
se <- airway
vizection(se)


# Older way to run Vizection:
data(iris)
genes <- iris[, 1:4]
libs <- as.data.frame(iris[, "Species"])
colnames(libs) <- c('group')
libs$samplename <- as.character(1:nrow(genes))
rownames(libs) <- libs$samplename 
libs$counts <- rnorm(n = nrow(genes), mean = 1000, sd = 200)
genes <- as.data.frame(t(genes))
vizection(genes = genes, libs = libs, local = FALSE)
}
}
