% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizectionServerFunctions.R
\name{vizectionExampleEnv}
\alias{vizectionExampleEnv}
\title{vizectionExampleEnv}
\usage{
vizectionExampleEnv()
}
\description{
Toy example of "input" object passed by Shiny
}
\details{
This is for debugging purposes; the example does not reflect well
the kind of real data that is expected.
}
